Integrity check for Eliza renormalization sequence number
