+++ kifi-backend/modules/eliza/app/com/keepit/eliza/commanders/ElizaStatsCommander.scala
@@ -1,14 +1,15 @@
 package com.keepit.eliza.commanders
 
 import com.google.inject.Inject
-import com.keepit.eliza.model.{UserThreadStats, UserThreadRepo}
+import com.keepit.eliza.model.{UriRenormalizationTrackingRepo, UserThreadStats, UserThreadRepo}
 import com.keepit.common.db.slick.Database
 import com.keepit.common.logging.Logging
-import com.keepit.common.db.Id
+import com.keepit.common.db.{SequenceNumber, Id}
 import com.keepit.model.User
 
 class ElizaStatsCommander @Inject() (
   userThreadRepo: UserThreadRepo,
+  renormalizationRepo: UriRenormalizationTrackingRepo,
   db: Database) extends Logging {
 
   def getUserThreadStats(userId: Id[User]): UserThreadStats = {
@@ -16,4 +17,7 @@ class ElizaStatsCommander @Inject() (
       userThreadRepo.getUserStats(userId)
     }
   }
-}
+
+  def getCurrentRenormalizationSequenceNumber(): Long = db.readOnly { implicit session => renormalizationRepo.getCurrentSequenceNumber() }
+
+  }
