+++ kifi-backend/modules/eliza/app/com/keepit/eliza/controllers/internal/ElizaController.scala
@@ -11,7 +11,7 @@ import scala.concurrent.future
 import play.api.libs.concurrent.Execution.Implicits.defaultContext
 
 import play.api.mvc.Action
-import play.api.libs.json.{JsObject, JsArray}
+import play.api.libs.json.{JsNumber, JsObject, JsArray}
 
 import com.google.inject.Inject
 import com.keepit.eliza.commanders.ElizaStatsCommander
@@ -58,4 +58,8 @@ class ElizaController @Inject() (
     Ok(notificationRouter.connectedSockets.toString)
   }
 
+  def getRenormalizationSequenceNumber() = Action { _ =>
+    val seqNumber = elizaStatsCommander.getCurrentRenormalizationSequenceNumber
+    Ok(JsNumber(seqNumber))
+  }
 }
