+++ kifi-backend/modules/eliza/conf/eliza.routes
@@ -36,6 +36,7 @@ POST  /internal/eliza/sendToUser                @com.keepit.eliza.controllers.in
 POST  /internal/eliza/sendToAllUsers            @com.keepit.eliza.controllers.internal.ElizaController.sendToAllUsers
 GET   /internal/eliza/connectedClientCount      @com.keepit.eliza.controllers.internal.ElizaController.connectedClientCount
 GET   /internal/eliza/getUserThreadStats        @com.keepit.eliza.controllers.internal.ElizaController.getUserThreadStats(userId: Id[User])
+GET   /internal/eliza/sequenceNumber/renormalization @com.keepit.eliza.controllers.internal.ElizaController.getRenormalizationSequenceNumber()
 
 POST  /internal/eliza/sendGlobalNotification    @com.keepit.eliza.controllers.internal.MessagingController.sendGlobalNotification
 GET   /internal/eliza/getThreadContentForIndexing  @com.keepit.eliza.controllers.internal.MessagingController.getThreadContentForIndexing(sequenceNumber: Long, maxBatchSize: Long)
@@ -43,6 +44,7 @@ GET   /internal/eliza/getThreadContentForIndexing  @com.keepit.eliza.controllers
 GET   /internal/eliza/verifyAllNotifications    @com.keepit.eliza.controllers.internal.MessagingController.verifyAllNotifications
 
 GET   /internal/eliza/showMeSomeEmail           @com.keepit.eliza.controllers.internal.MessagingController.showMeSomeEmail()
+
 
 ##########################################
 # IMPORTS
