+++ kifi-backend/modules/shoebox/app/com/keepit/integrity/SequenceNumberChecker.scala
@@ -1,0 +1,47 @@
+package com.keepit.integrity
+
+import com.google.inject.Inject
+import com.keepit.common.db.slick.Database
+import com.keepit.model.ChangedURIRepo
+import com.keepit.eliza.ElizaServiceClient
+import com.keepit.common.healthcheck.{AirbrakeError, AirbrakeNotifier}
+import com.keepit.common.service.ServiceType
+import play.api.libs.concurrent.Execution.Implicits.defaultContext
+import com.keepit.common.logging.Logging
+
+trait SequenceNumberChecker {
+  val service: ServiceType
+  def check(): Unit
+}
+
+class SequenceNumberOffException(message: String) extends Exception(message)
+
+class ElizaSequenceNumberChecker @Inject() (
+  db: Database,
+  changedUriRepo: ChangedURIRepo,
+  elizaServiceClient: ElizaServiceClient,
+  airbrake: AirbrakeNotifier
+) extends SequenceNumberChecker with Logging {
+
+  private val threshold = 100
+  val service = ServiceType.ELIZA
+
+  def check(): Unit = {
+    checkRenormalizationSequenceNumber()
+  }
+
+  protected def checkRenormalizationSequenceNumber(): Unit = {
+    val elizaRenormalizationSequenceNumber = elizaServiceClient.getRenormalizationSequenceNumber()
+    val shoeboxRenormalizationSequenceNumber = db.readOnly { implicit session => changedUriRepo.getHighestSeqNum().get.value }
+    elizaRenormalizationSequenceNumber.foreach { elizaSeq =>
+      log.info(s"[Renormalization] Sequence Numbers: Shoebox -> $shoeboxRenormalizationSequenceNumber, Eliza -> $elizaSeq")
+      if (shoeboxRenormalizationSequenceNumber - elizaSeq > threshold) {
+        airbrake.notify(AirbrakeError(new SequenceNumberOffException(s"[Renormalization] Eliza is falling behind, at sequence number $elizaSeq while Shoebox is at $shoeboxRenormalizationSequenceNumber")))
+      }
+    }
+  }
+}
+
+
+
+
